requirements.txt
discord.py
aiohttp
azure-identity
azure-mgmt-compute